Make price more prominent than "free trial".
According to Apple reject reason, the "free trial" part must be less prominent than the price. The mention of "free trial" is now removed from the top title and text. A new label near the price has the trial mention in a non-bold font, same size as the price label which is now in a bold font.