Implement new tracking methods
calendar_subscription: This event replaces subscribe to be more clear of what we are tracking, this event is triggered when a user finish subscribing to a calendar and the calendar is added to the user’s device.

show_purchase_screen: This is a new event to better understand the user’s reaction to the purchase screen, this event is triggered when the purchase screen (blue one with the price) appears.

click_purchase_button: This new event will trigger when the user taps on the purchase button at the bottom of the purchase screen.

subscription_purchase:  This new event will trigger when the in-app purchase of a subscription to our service is successful.

subscription_failed : This new event will trigger when the in-app purchase of a subscription to our service failed to be processed by apple.